refactor: 첫번째 상행역을 찾는 메소드 Sections 로 이동
